(minor) added a killswitch and some debug settings
Adds a killswitch command (kill) to stop the bot manually when working on it - no need to go onto the server to stop it anymore.  Needs botOwnerID in config.json.

added a debug switch - config.json needs to be reconfigured for this so you can turn it on or off.